backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
face_recognition
python-dotenv
google-cloud-firestore